requirements.txt
Flask==3.1.0
langchain_community==0.3.21
langchain_core==0.3.51
langchain_openai==0.3.12
langgraph==0.3.30
python-dotenv==1.1.0
Werkzeug==3.1.3
gunicorn